Add files via upload
In this study, PX-Net, an automated segmentation model for the segmentation of  the renal cortex, and the renal medulla using computed tomography (CT) images was proposed with deep learning (DL) technique.
In this repository, A sample CT image was given as as an input, and  code have been made publicly available for deep learning studies. You can run the given code using Matlab 2021a.